Locate ewk_view_page_close in extended main loop
Sometimes, closing page callbkack is missing when the callback comes
between terminating main loop and creating extended main loop.
In that case, deadlock situation happens.
To avoid the issue, the ewk_view_page_close needs to be called
in extended main loop to assure receiving the callback.
Also, evas object needs to be deleted when it's valid.

Bug: http://suprem.sec.samsung.net/jira/browse/TWF-2782

Change-Id: I7a86121bc7c78225b54bb6fe22b934f62615a46d
Signed-off-by: Youngsoo Choi <kenshin.choi@samsung.com>